Create the new Information Technology Law
This Law addresses the need to present a unified stance of Liberland towards information systems. The point is to codify how Liberland uses systems of electronic communication. This use can create dependence, particularly on one given developer. That developer can then start providing suboptimal services for extraordinary prices. The cure for that is the reliance to open source and free software. Also, the institutional setup of technical agencies must be addressed.