Criminal conviction information
Criminal convictions are special information that the states wish to monopolize access to. To become compliant, we accept this wish. In short: We will only access officially posted State registers to retrieve information on convictions. No hearsay or secondary sources are permissible. Also, if the State party doesn't provide sufficient safeguards, access shall likewise not be permissible.